Replace @reach/utils with react-use. Clean up _app.tsx. Remove hydration warning.
