+++ features/liquidity/components/BondLiquidityDialog.tsx
@@ -1,4 +1,3 @@
-import { useUpdateEffect } from '@reach/utils'
 import dayjs from 'dayjs'
 import { useBondTokens, useUnbondTokens } from 'hooks/useBondTokens'
 import { useRefetchQueries } from 'hooks/useRefetchQueries'
@@ -24,6 +23,7 @@ import { useQueryPoolLiquidity } from 'queries/useQueryPools'
 import { useQueryPoolUnstakingDuration } from 'queries/useQueryPoolUnstakingDuration'
 import React, { useEffect, useRef, useState } from 'react'
 import { toast } from 'react-hot-toast'
+import { useUpdateEffect } from 'react-use'
 import { formatSdkErrorMessage } from 'util/formatSdkErrorMessage'
 
 import { LiquidityInputSelector } from './LiquidityInputSelector'
