+++ package.json
@@ -37,6 +37,7 @@
     "react-dom": "^18.2.0",
     "react-hot-toast": "^2.2.0",
     "react-query": "^3.39.1",
+    "react-use": "^17.4.0",
     "recoil": "^0.7.4",
     "sass": "^1.45.0"
   },
