+++ pages/_app.tsx
@@ -4,29 +4,15 @@ import 'focus-visible'
 
 import { ErrorBoundary } from 'components/ErrorBoundary'
 import { TestnetDialog } from 'components/TestnetDialog'
-import {
-  css,
-  globalCss,
-  media,
-  styled,
-  useMedia,
-  useSubscribeDefaultAppTheme,
-  useThemeClassName,
-} from 'junoblocks'
+import { css, media, useMedia } from 'junoblocks'
 import type { AppProps } from 'next/app'
-import { useEffect } from 'react'
 import { Toaster } from 'react-hot-toast'
 import { QueryClientProvider } from 'react-query'
 import { RecoilRoot } from 'recoil'
 import { queryClient } from 'services/queryClient'
 
+import { NextJsAppRoot } from '../components/NextJsAppRoot'
 import { __TEST_MODE__ } from '../util/constants'
-
-const applyGlobalStyles = globalCss({
-  body: {
-    backgroundColor: '$backgroundColors$base',
-  },
-})
 
 const toasterClassName = css({
   [media.sm]: {
@@ -35,36 +21,6 @@ const toasterClassName = css({
     bottom: '$6 !important',
   },
 }).toString()
-
-function NextJsAppRoot({ children }) {
-  const themeClassName = useThemeClassName()
-
-  useSubscribeDefaultAppTheme()
-
-  /* apply theme class on body also */
-  useEffect(() => {
-    document.body.classList.add(themeClassName)
-    applyGlobalStyles()
-    return () => {
-      document.body.classList.remove(themeClassName)
-    }
-  }, [themeClassName])
-
-  return (
-    <StyledContentWrapper
-      data-app-wrapper=""
-      lang="en-US"
-      className={typeof window === 'undefined' ? null : themeClassName}
-      suppressHydrationWarning
-    >
-      {typeof window === 'undefined' ? null : children}
-    </StyledContentWrapper>
-  )
-}
-
-const StyledContentWrapper = styled('div', {
-  backgroundColor: '$backgroundColors$base',
-})
 
 function MyApp({ Component, pageProps }: AppProps) {
   const isSmallScreen = useMedia('sm')
